+++ R/utils.R
@@ -5,14 +5,25 @@ scale_runif <- function(x, new_min, new_max) {
 }
 
 make_options <- function(
-  cvd = c("protan", "deutan", "tritan"),
-  cvd_severity = 0,
+  cvd = c(protan = 0, deutan = 0, tritan = 0),
+  cvd_severity,
   bg = NULL,
   metric = c("din99d", "ciede2000", "cie76"),
   extend = NULL
 ) {
-  cvd <- match.arg(cvd)
   metric <- match.arg(metric)
+
+  if (!missing(cvd_severity)) {
+    message(
+      "The argument 'cvd_severity' is deprecated and will be
+      removed in a future version. Please use a named 'cvd' object instead."
+    )
+    stopifnot(
+      length(cvd_severity) == 1,
+      cvd_severity >= 0,
+      cvd_severity <= 1
+    )
+  }
 
   if (is.null(extend)) {
     extend <- matrix(0, nrow = 0, ncol = 3)
@@ -24,19 +35,32 @@ make_options <- function(
   }
 
   stopifnot(
-    is.character(cvd),
-    length(cvd) == 1,
-    is.numeric(cvd_severity),
-    length(cvd_severity) == 1,
-    cvd_severity >= 0,
-    cvd_severity <= 1,
+    is.character(cvd) || is.numeric(cvd) || is.null(cvd),
     is.null(bg) || (is.character(bg) && length(bg) == 1),
     is.matrix(extend),
     ncol(extend) == 3
   )
 
-  cvd_list <- list()
-  cvd_list[[cvd]] <- cvd_severity
+  if (is.null(cvd)) {
+    cvd_list <- list(protan = 0, deutan = 0, tritan = 0)
+  } else if (is.character(cvd)) {
+    cvd <- match.arg(cvd, c("protan", "deutan", "tritan"))
+    cvd_list <- list()
+    cvd_list[[cvd]] <- cvd_severity
+  } else {
+    stopifnot(
+      is.numeric(cvd),
+      length(cvd) <= 3,
+      all(cvd >= 0),
+      all(cvd <= 1),
+      names(cvd) %in% c("protan", "deutan", "tritan")
+    )
+    cvd_list <- list(
+      protan = ifelse("protan" %in% names(cvd), cvd["protan"], 0),
+      deutan = ifelse("deutan" %in% names(cvd), cvd["deutan"], 0),
+      tritan = ifelse("tritan" %in% names(cvd), cvd["tritan"], 0)
+    )
+  }
 
   if (!is.null(bg)) {
     bg <- grDevices::col2rgb(bg)